+++ app/lib/features/connection/domain/supabase_config.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/foundation.dart';
+
+@immutable
+class SupabaseConfig {
+  const SupabaseConfig({required this.url, required this.anonKey});
+
+  final String url;
+  final String anonKey;
+}